Fix: Shop Endpoint Refactor Complete
