+++ super_admin_1/shop/del_shop.py
@@ -9,39 +9,11 @@ from super_admin_1.shop.shoplog_helpers import ShopLogs
 import os
 from utils import super_admin_required
 
-del_shop = Blueprint('del_shop', __name__, url_prefix='/api/shop')
-
-
-
-@del_shop.route('/<shop_id>', methods=['PATCH'], strict_slashes=False)
-@super_admin_required
-def delete_shop(shop_id):
-    """Delete a shop"""
-    # verify if shop exists
-    shop = Shop.query.filter_by(id=shop_id).first()
-    if not shop:
-        return jsonify({'forbidden': 'Shop not found'}), 404
-    # check if shop is temporary
-    if shop.is_deleted == 'temporary':
-        return jsonify({'message': 'Shop already deleted'}), 400
-    # delete shop temporarily
-    shop.is_deleted = 'temporary'
-    db.session.commit()
-
-    """
-    The following logs the action in the shop_log db
-    """
-    get_user_id = shop.user.id
-    action = ShopLogs(
-        shop_id=shop_id,
-        user_id=get_user_id
-    )
-    action.log_shop_deleted(delete_type="temporary")
-    return jsonify({'message': 'Shop temporarily deleted'}), 200
+test = Blueprint('test', __name__, url_prefix='/api/test')
 
 
 # ============================== MY HELPER FUNCTON ================================
-@del_shop.route('/user/create', methods=['POST'])
+@test.route('/user/create', methods=['POST'])
 @super_admin_required
 def create_user():
     """ Create a new user"""
@@ -58,7 +30,7 @@ def create_user():
     return jsonify(user.format()), 201
 
 
-@del_shop.route('/user/<user_id>/shop', methods=['POST'])
+@test.route('/user/<user_id>/shop', methods=['POST'])
 @super_admin_required
 def create_shop(user_id):
     """ Create a new shop"""
@@ -89,8 +61,8 @@ def create_shop(user_id):
 
 
 # get request for shop
-@del_shop.route('/', methods=['GET'], strict_slashes=False, defaults={'shop_id': None})
-@del_shop.route('/<shop_id>', methods=['GET'])
+@test.route('/', methods=['GET'], strict_slashes=False, defaults={'shop_id': None})
+@test.route('/<shop_id>', methods=['GET'])
 @super_admin_required
 def get_shop(shop_id):
     """ Get a shop or all shop"""
@@ -100,22 +72,10 @@ def get_shop(shop_id):
         return jsonify([shop.format() for shop in Shop.query.all()]), 200
 
 
-# delete shop object permanently out of the DB
-@del_shop.route('/<shop_id>/del', methods=['DELETE'])
-@super_admin_required
-def perm_del(shop_id):
-    """ Delete a shop"""
-    shop = Shop.query.filter_by(id=shop_id).first()
-    if not shop:
-        abort(404)
-    db.session.delete(shop)
-    db.session.commit()
-    return jsonify({'message': 'Shop deleted aggresively'}), 200
-
 
 # get request for user
-@del_shop.route('/user', methods=['GET'], strict_slashes=False, defaults={'user_id': None})
-@del_shop.route('/user/<user_id>', methods=['GET'])
+@test.route('/user', methods=['GET'], strict_slashes=False, defaults={'user_id': None})
+@test.route('/user/<user_id>', methods=['GET'])
 @super_admin_required
 def get_user(user_id):
     """ Get all user"""
@@ -124,10 +84,9 @@ def get_user(user_id):
     else:
         return jsonify([user.format() for user in User.query.all()]), 200
 
+
 # delete user object
-
-
-@del_shop.route('/user/<user_id>', methods=['DELETE'])
+@test.route('/user/<user_id>', methods=['DELETE'])
 @super_admin_required
 def delete_user(user_id):
     """ Delete a user"""
@@ -140,59 +99,3 @@ def delete_user(user_id):
 # ======================================== HELPER FUNCTIN END=============================================
 
 
-@del_shop.route("/logs/shops", defaults={"shop_id": None})
-@del_shop.route("/logs/shops/<int:shop_id>")
-@super_admin_required
-def get_all_shop_logs(shop_id):
-    """Get all shop logs"""
-    if not shop_id:
-        return (
-            jsonify(
-                {
-                    "message": "success",
-                    "logs": [
-                        log.format() if log else [] for log in ShopsLogs.query.all()
-                    ],
-                }
-            ),
-            200,
-        )
-
-    return (
-        jsonify(
-            {
-                "message": "success",
-                "logs": [
-                    log.format() if log else []
-                    for log in ShopsLogs.query.filter_by(shop_id=shop_id).all()
-                ],
-            }
-        ),
-        200,
-    )
-
-
-@del_shop.route("/logs/shops/download", defaults={"shop_id": None})
-@del_shop.route("/logs/shops/<int:shop_id>/download")
-@super_admin_required
-def download_shop_logs(shop_id):
-    """Download all shop logs"""
-    logs = []
-    if not shop_id:
-        logs = [log.format() if log else [] for log in ShopsLogs.query.all()]
-    else:
-        logs = [
-            log.format() if log else []
-            for log in ShopsLogs.query.filter_by(shop_id=shop_id).all()
-        ]
-    # Create a temporary file to store the strings
-    temp_file_path = f"{os.path.abspath('.')}/temp_file.txt"
-    with open(temp_file_path, "w") as temp_file:
-        temp_file.write("\n".join(logs))
-
-    response = send_file(
-        temp_file_path, as_attachment=True, download_name="shoplogs.txt"
-    )
-    os.remove(temp_file_path)
-
-    return response
