--- super_admin_1/shop/shop_activity.py
@@ -1,14 +1,0 @@
-from super_admin_1 import db
-from flask import Blueprint, jsonify
-from super_admin_1.models.shop_logs import ShopsLogs
-from utils import super_admin_required
-
-
-events = Blueprint("events", __name__, url_prefix="/api/events")
-
-
-@events.route("/shop/actions", methods=["GET"])
-@super_admin_required
-def shop_actions():
-    data = ShopsLogs.query.all()
-    return jsonify([action.format_json() for action in data]), 200
